// gui/frontend/src/store/redux-query/models/QueueParam.ts
// tslint:disable
/**
 * Onyks Wagon Location API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import {
  Queue,
  QueueFromJSON,
  QueueToJSON,
  QueueParamDef,
  QueueParamDefFromJSON,
  QueueParamDefToJSON,
} from ".";

/**
 *
 * @export
 * @interface QueueParam
 */
export interface QueueParam {
  /**
   *
   * @type {string}
   * @memberof QueueParam
   */
  paramsDefId: string;
  /**
   *
   * @type {string}
   * @memberof QueueParam
   */
  value?: string;
  /**
   *
   * @type {Queue}
   * @memberof QueueParam
   */
  module?: Queue;
  /**
   *
   * @type {QueueParamDef}
   * @memberof QueueParam
   */
  paramDef?: QueueParamDef;
}

export function QueueParamFromJSON(json: any): QueueParam {
  return {
    paramsDefId: json["params_def_id"],
    value: !exists(json, "value") ? undefined : json["value"],
    module: !exists(json, "module") ? undefined : QueueFromJSON(json["module"]),
    paramDef: !exists(json, "param_def")
      ? undefined
      : QueueParamDefFromJSON(json["param_def"]),
  };
}

export function QueueParamToJSON(value?: QueueParam): any {
  if (value === undefined) {
    return undefined;
  }
  return {
    params_def_id: value.paramsDefId,
    value: value.value,
    module: QueueToJSON(value.module),
    param_def: QueueParamDefToJSON(value.paramDef),
  };
}